+++ lib/src/application/deposit_usecase.dart
@@ -1,3 +1,5 @@
+import 'package:sikka_pay/src/domain/entities/transaction.dart';
+
 import '../infrastructure/account_repository.dart';
 
 class DepositUseCase {
@@ -13,5 +15,14 @@ class DepositUseCase {
     final account = repository.getAccountOrThrow(phoneNumber);
     account.balance += amount;
     repository.saveAccount(account);
+
+    repository.saveTransaction(
+      Transaction(
+        phoneNumber: phoneNumber,
+        type: TransactionType.deposit,
+        amount: amount,
+        date: DateTime.now(),
+      ),
+    );
   }
 }