+++ lib/src/application/withdraw_usecase.dart
@@ -1,3 +1,5 @@
+import 'package:sikka_pay/src/domain/entities/transaction.dart';
+
 import '../infrastructure/account_repository.dart';
 import '../domain/exceptions/insufficient_funds_exception.dart';
 
@@ -23,5 +25,14 @@ class WithdrawUseCase {
 
     account.balance -= amount;
     repository.saveAccount(account);
+
+    repository.saveTransaction(
+      Transaction(
+        phoneNumber: phoneNumber,
+        type: TransactionType.withdrawal,
+        amount: amount,
+        date: DateTime.now(),
+      ),
+    );
   }
 }