+++ lib/src/domain/datasources/account_data_source.dart
@@ -1,3 +1,5 @@
+import 'package:sikka_pay/src/domain/entities/transaction.dart';
+
 import '../entities/account.dart';
 
 abstract class AccountDataSource {
@@ -6,4 +8,6 @@ abstract class AccountDataSource {
   bool exists(String phoneNumber);
   bool authenticate(String phoneNumber, String pinCode);
   void clear(); // Pour les tests
+  void saveTransaction(Transaction tx);
+  List<Transaction> getTransactions(String phoneNumber);
 }