+++ lib/src/infrastructure/fake_account_data_source.dart
@@ -1,8 +1,11 @@
+import 'package:sikka_pay/src/domain/entities/transaction.dart';
+
 import '../domain/datasources/account_data_source.dart';
 import '../domain/entities/account.dart';
 
 class FakeAccountDataSource implements AccountDataSource {
   final Map<String, Account> _accounts = {};
+  final Map<String, List<Transaction>> _transactionHistory = {};
 
   @override
   Account? getAccount(String phoneNumber) => _accounts[phoneNumber];
@@ -23,4 +26,15 @@ class FakeAccountDataSource implements AccountDataSource {
 
   @override
   void clear() => _accounts.clear();
+
+  @override
+  void saveTransaction(Transaction tx) {
+    _transactionHistory.putIfAbsent(tx.phoneNumber, () => []);
+    _transactionHistory[tx.phoneNumber]!.add(tx);
+  }
+
+  @override
+  List<Transaction> getTransactions(String phoneNumber) {
+    return _transactionHistory[phoneNumber] ?? [];
+  }
 }